Moved provenance.py functions to cwlprov/__init__.py
